+++ CHANGELOG.md
@@ -30,7 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Windows MSBuild executables have the wrong version #1745
 - Cast Overflow at pango_textlayout #1314
 - x11 back end segfaults if display is unavailable #1776
-- Repeated file read gives different results with libcgraph #1767
 - typo in cmd/gvpr/lib/clustg #1781
 - Segfault in dot #1783
 - Incorrect 'Arrow type "s" unknown' error #1444
@@ -72,6 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - gml2gv doesn't handle some attributes correctly #1869
 - Add missing circo, fdp, neato, osage, patchwork, sfdp & twopi tools to Windows builds (copies of dot)
 - Add gv2gml tool to CMake (copy of gml2gv on Windows, symlink to gml2gv otherwise)
+- Regression: fdp generates internal names in the output #1876
+- Regression: fdp assertion error on cluster in edge #1877
 
 ## [2.44.1] - 2020-06-29